BUG/MAJOR: reload: Rework of the reload agent
There are small number of cases where reload will fail when -c
passes (e.g. cannot bind socket from bind line), in that case
reload would fail and haproxy crashed. To deal with these cases
added a new CLI option -s (--restart-cmd), so now when reload
fails dataplaneapi restarts haproxy using this parameter with the
last known good config file.

It assumes that the config file given on the start is valid and
copies it with .lkg suffix. On every successful reload that .lkg
file is replaced with a working cfg. When the reload fails, it
replases the <--config-file> with <--config-file.lkg> restarts the
HAProxy and returns the bogus <--config-file>.