MEDIUM: runtime: Rework runtime client inits
Dataplaneapi checks if master socket is configured and uses only
master socket for Runtime API in combination with configured
nbproc.

If no master socket is configured and nbproc=1, we read the first
valid unix socket and use that for Runtime API.

If no master socket is configured and nbproc>1, we look for stats
socket lines with configured process keyword (we ignore all, odd,
even and ranged values) and use those sockets for the respective
processes. If no such socket is configured we do not set up a
Runtime API. If not all processes have a stats socket each, we emit
a warning and continue.