BUG/MAJOR: discovery: Add type assertion check to DiscoverChildPaths
A panic is caused when DiscoverChildPaths loops over endpoints that
do not have a "get" method defined within embedded_spec.go. At time of
this commit it is caused specifically by "/cluster/certificate" only
having "post" and not a "get" defined. This commit adds a type assertion
check to verify that v["get"] is of type map[string]interface{}.

This fixes #85