BUG: Prevent a panic in the case when a map only has 1 entry
When a map only has a single entry `rand.Intn()` is supplied with `0` as
as parameter which is invalid.

In addition the previous code included a `+ 1` to the result of `rand.Intn()`.
This is not needed, `rand.Intn()` provides results in the range of 0-(n-1)
which will align with the array indexes.