MINOR: auth: skipping basic authentication in case of mTLS
The Basic Authentication can be disabled when Dataplane API is running
in TLS/SSL and the Certificate Authority is defined using the backing
environment variable `TLS_CA_CERTIFICATE`.

If this precondition is not met, TLS/SSL will be enabled but the Basic
Authentication will be still required since the client certificate
requirement and verification would not be possible.