MEDIUM: transaction: marking outdated transactions on commit
When dealing with multiple transactions the end-user could end up with
versions that are not anymore committable due to mismatching version
number.

This commit aims to perform the clean-up process in a fashionated
automated way without putting this task to the end-user.

Outdated transactions will be placed in the /tmp/haproxy/outdated folder
for further investigation.