MEDIUM: transaction: considering only in_progress files
According to changes introduced by a5c1fef we cannot anymore rely on the
Increase or Decrease functionalities of the Threshold interface since
the clean-up upon transaction commit cannot be propagated from the
bottom to the up.

The change is accepting a closure that will return the actual
`in_progres` transactions, simplifying the code with the need of
decorating just the `transactions.StartTransactionHandler` handler.