MINOR: discovery: Add debug output for AWS service discovery
This patch adds `log.Debugf()` statements to a few key places
within the AWS service discovery code. Specifically, the starting
and stopping points for the service discovery job, places where it
is stopped due to errors, when services are added.

Being that service discovery is something that automagically happens,
it can be helpful to have a reference point as to where or why something
is going wrong.