BUG/MINOR: discovery: AWS handle version mismatch
When multiple AWS regions/clusters are specified and the
HAProxy Data Plane API attempts to update them at the same time it
is possible that one of the transactions is still in process when
another transaction is initiated. This causes the version ID to not
have been incremented yet, which can cause a version mismatch error.
This patch catches the configuration.ErrVersionMismatch error and
allows the transaction to be attempted again, which allows it to
get the latest version.

This fixes #194.