BUG/MINOR: detect when a raw configuration is uploaded incorrectly
Return an error when the user tries to upload an HAProxy configuration
file without any newline character. This can happen for example when
using curl(1) with bad options:

  # BAD! This will url-encode the file.
  curl -H 'Content-Type: text/plain' -d @haproxy.cfg $URL

  # GOOD. Upload without encoding.
  curl -H 'Content-Type: text/plain' -T haproxy.cfg $URL

Note that we still allow to upload an empty configuration, since this
is supported by HAProxy in master-worker mode, and it may actually be
useful in future releases.