BUG/MINOR: fix panic when trying to start with hcl configuration file
hcl format is not supported anymore. Only yaml format is supported. The problem was that storage was not initialized if the file format was not yaml.
I kept the same behavior. If the file extension is not 'yaml' or 'yml', we still try to parse the configuration as a yaml file.

Signed-off-by: Vincent Gramer <vgramer@haproxy.com>