BUG/MAJOR: reconfigure runtime client and HAPorxy if path changes in the configuration file
If the user manually changes the configuration, the file watcher will be notified, the configuration will be reloaded, and the runtime client will be reconfigured if needed.
To be able to reconfigure the runtime client, HAProxy must be reloaded. Otherwise, the new socket does not exist yet, and reconfiguration fails. We use the Reload agent to schedule a reload and reconfigure the runtime client.

Signed-off-by: Vincent Gramer <vgramer@haproxy.com>