MEDIUM: API consolidation- add PUT on index based lists and remove index in objects
This MR adds a PUT on all index based lists in order to allow to replace all elements at once.
It also removes the index that were in objects definition and adds the index to the URL. This will change for POST one element that was before POST /acls and is now POST /acls/{index}.
The list based have now:
- POST /list_based with the body having a list of objects.
They will be inserted in the same order as sent in the list.

Also: MINOR: remove log_message in tcp_check and use only check_comment